feat: create ActivitySelectClientPage component with tab navigation and tour functionality; integrate dynamic import for client-side rendering
